Fixes STR#3041, int -> size_t
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@10086 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Native_File_Chooser_WIN32.cxx
@@ -710,7 +710,7 @@ void Fl_Native_File_Chooser::parse_filter(const char *in) {
 
   // whatever input string is, our output won't be much longer in length..
   // use double length just for safety.
-  int slen = strlen(in);
+  size_t slen = strlen(in);
   char *wildprefix = new char[slen*2]; wildprefix[0] = 0;
   char *comp       = new char[slen*2]; comp[0] = 0;
   char *name       = new char[slen*2]; name[0] = 0;
